+++ engine/table/src/main/java/io/deephaven/engine/table/impl/updateby/em/BigIntegerEMOperator.java
@@ -107,7 +107,7 @@ public class BigIntegerEMOperator extends BaseBigNumberEMOperator<BigInteger> {
                                 @Nullable final RowRedirection rowRedirection,
                                 @NotNull final OperationControl control,
                                 @Nullable final String timestampColumnName,
-                                final long windowScaleUnits,
+                                final double windowScaleUnits,
                                 final ColumnSource<?> valueSource,
                                 @NotNull final EmFunction aggFunction) {
         super(pair, affectingColumns, rowRedirection, control, timestampColumnName, windowScaleUnits, valueSource, aggFunction);
