+++ engine/table/src/main/java/io/deephaven/engine/table/impl/updateby/em/ByteEMOperator.java
@@ -83,7 +83,7 @@ public class ByteEMOperator extends BasePrimitiveEMOperator {
                         final long dt = timestamp - lastStamp;
                         if (dt != lastDt) {
                             // Alpha is dynamic based on time, but only recalculated when needed
-                            alpha = Math.exp(-dt / (double) reverseWindowScaleUnits);
+                            alpha = Math.exp(-dt / reverseWindowScaleUnits);
                             oneMinusAlpha = 1.0 - alpha;
                             lastDt = dt;
                         }
@@ -125,7 +125,7 @@ public class ByteEMOperator extends BasePrimitiveEMOperator {
                           @Nullable final RowRedirection rowRedirection,
                           @NotNull final OperationControl control,
                           @Nullable final String timestampColumnName,
-                          final long windowScaleUnits,
+                          final double windowScaleUnits,
                           final ColumnSource<?> valueSource,
                           @NotNull final EmFunction aggFunction
                           // region extra-constructor-args
