+++ engine/table/src/main/java/io/deephaven/engine/table/impl/updateby/em/CharEMOperator.java
@@ -77,7 +77,7 @@ public class CharEMOperator extends BasePrimitiveEMOperator {
                         final long dt = timestamp - lastStamp;
                         if (dt != lastDt) {
                             // Alpha is dynamic based on time, but only recalculated when needed
-                            alpha = Math.exp(-dt / (double) reverseWindowScaleUnits);
+                            alpha = Math.exp(-dt / reverseWindowScaleUnits);
                             oneMinusAlpha = 1.0 - alpha;
                             lastDt = dt;
                         }
@@ -119,7 +119,7 @@ public class CharEMOperator extends BasePrimitiveEMOperator {
                           @Nullable final RowRedirection rowRedirection,
                           @NotNull final OperationControl control,
                           @Nullable final String timestampColumnName,
-                          final long windowScaleUnits,
+                          final double windowScaleUnits,
                           final ColumnSource<?> valueSource,
                           @NotNull final EmFunction aggFunction
                           // region extra-constructor-args
