+++ engine/table/src/main/java/io/deephaven/engine/table/impl/updateby/em/DoubleEMOperator.java
@@ -79,7 +79,7 @@ public class DoubleEMOperator extends BasePrimitiveEMOperator {
                     } else {
                         final long dt = timestamp - lastStamp;
                         if (dt != 0) {
-                            final double alpha = Math.exp(-dt / (double) reverseWindowScaleUnits);
+                            final double alpha = Math.exp(-dt / reverseWindowScaleUnits);
                             final double oneMinusAlpha = 1.0 - alpha;
                             curVal = aggFunction.apply(curVal, input, alpha, oneMinusAlpha);
                             lastStamp = timestamp;
@@ -130,7 +130,7 @@ public class DoubleEMOperator extends BasePrimitiveEMOperator {
                            @Nullable final RowRedirection rowRedirection,
                            @NotNull final OperationControl control,
                            @Nullable final String timestampColumnName,
-                           final long windowScaleUnits,
+                           final double windowScaleUnits,
                            final ColumnSource<?> valueSource,
                            @NotNull final EmFunction aggFunction
                            // region extra-constructor-args
