+++ engine/table/src/main/java/io/deephaven/engine/table/impl/updateby/em/LongEMOperator.java
@@ -82,7 +82,7 @@ public class LongEMOperator extends BasePrimitiveEMOperator {
                         final long dt = timestamp - lastStamp;
                         if (dt != lastDt) {
                             // Alpha is dynamic based on time, but only recalculated when needed
-                            alpha = Math.exp(-dt / (double) reverseWindowScaleUnits);
+                            alpha = Math.exp(-dt / reverseWindowScaleUnits);
                             oneMinusAlpha = 1.0 - alpha;
                             lastDt = dt;
                         }
@@ -124,7 +124,7 @@ public class LongEMOperator extends BasePrimitiveEMOperator {
                           @Nullable final RowRedirection rowRedirection,
                           @NotNull final OperationControl control,
                           @Nullable final String timestampColumnName,
-                          final long windowScaleUnits,
+                          final double windowScaleUnits,
                           final ColumnSource<?> valueSource,
                           @NotNull final EmFunction aggFunction
                           // region extra-constructor-args
